+++ bridgetown-core/lib/bridgetown-core/drops/resource_drop.rb
@@ -14,6 +14,7 @@ module Bridgetown
       def_delegator :@obj, :relative_path, :path
       def_delegators :@obj,
                      :id,
+                     :data,
                      :output,
                      :content,
                      :to_s,
@@ -21,7 +22,8 @@ module Bridgetown
                      :relative_path,
                      :relative_url,
                      :date,
-                     :taxonomies
+                     :taxonomies,
+                     :relations
 
       private def_delegator :@obj, :data, :fallback_data