+++ bridgetown-core/lib/bridgetown-core/helpers.rb
@@ -4,8 +4,13 @@ module Bridgetown
   class RubyTemplateView
     class Helpers
       include Bridgetown::Filters
+      include Bridgetown::Filters::FromLiquid
 
-      attr_reader :view, :site
+      # @return [Bridgetown::RubyTemplateView]
+      attr_reader :view
+
+      # @return [Bridgetown::Site]
+      attr_reader :site
 
       Context = Struct.new(:registers)
 
@@ -28,20 +33,44 @@ module Bridgetown
         pairs.select { |_key, truthy| truthy }.keys.join(" ")
       end
 
+      # Convert a Markdown string into HTML output.
+      #
+      # @param input [String] the Markdown to convert, if no block is passed
+      # @return [String]
+      def markdownify(input = nil, &block)
+        content = Bridgetown::Utils.reindent_for_markdown(
+          block.nil? ? input.to_s : view.capture(&block)
+        )
+        converter = site.find_converter_instance(Bridgetown::Converters::Markdown)
+        safe(converter.convert(content).strip)
+      end
+
       # This helper will generate the correct permalink URL for the file path.
       #
       # @param relative_path [String, Object] source file path, e.g.
-      #   "_posts/2020-10-20-my-post.md", or object that responds to `url`
+      #   "_posts/2020-10-20-my-post.md", or object that responds to either
+      #   `url` or `relative_url`
       # @return [String] the permalink URL for the file
-      # @raise [ArgumentError] if the file cannot be found
       def url_for(relative_path)
-        path_string = !relative_path.is_a?(String) ? relative_path.url : relative_path
+        if relative_path.respond_to?(:relative_url)
+          return safe(relative_path.relative_url) # new resource engine
+        elsif relative_path.respond_to?(:url)
+          return safe(relative_url(relative_path.url)) # old legacy engine
+        elsif relative_path.start_with?("/", "http")
+          return safe(relative_path)
+        end
 
-        return path_string if path_string.start_with?("/", "http")
+        find_relative_url_for_path(relative_path)
+      end
+      alias_method :link, :url_for
 
+      # @param relative_path [String] source file path, e.g.
+      #   "_posts/2020-10-20-my-post.md"
+      # @raise [ArgumentError] if the file cannot be found
+      def find_relative_url_for_path(relative_path)
         site.each_site_file do |item|
-          if item.relative_path == path_string || item.relative_path == "/#{path_string}"
-            return relative_url(item)
+          if item.relative_path == relative_path || item.relative_path == "/#{relative_path}"
+            return safe(item.respond_to?(:relative_url) ? item.relative_url : relative_url(item))
           end
         end
 
@@ -51,7 +80,6 @@ module Bridgetown
           Make sure the document exists and the path is correct.
         MSG
       end
-      alias_method :link, :url_for
 
       # This helper will generate the correct permalink URL for the file path.
       #
@@ -69,7 +97,8 @@ module Bridgetown
           attr = attr.to_s.tr("_", "-")
           segments << "#{attr}=\"#{Utils.xml_escape(option)}\""
         end
-        "<#{segments.join(" ")}>#{text}</a>"
+        # TODO: this might leak an XSS string into text, need to check
+        safe("<#{segments.join(" ")}>#{text}</a>")
       end
 
       # Forward all arguments to I18n.t method
@@ -79,6 +108,16 @@ module Bridgetown
       def t(*args)
         I18n.send :t, *args
       end
+
+      # For template contexts where ActiveSupport's output safety is loaded, we
+      # can ensure a string has been marked safe
+      #
+      # @param input [Object]
+      # @return [String]
+      def safe(input)
+        input.to_s.html_safe
+      end
+      alias_method :raw, :safe
     end
   end
 end