+++ bridgetown-core/lib/bridgetown-core/resource/taxonomy_type.rb
@@ -42,6 +42,15 @@ module Bridgetown
           "metadata" => metadata,
         }
       end
+      alias_method :to_h, :to_liquid
+
+      def as_json(*)
+        to_h
+      end
+
+      ruby2_keywords def to_json(*options)
+        as_json(*options).to_json(*options)
+      end
     end
   end
 end