+++ bridgetown-core/lib/bridgetown-core/resource/transformer.rb
@@ -12,18 +12,20 @@ module Bridgetown
       # @return [Bridgetown::Site]
       attr_reader :site
 
-      # @return [String]
-      attr_reader :output_ext
-
       def initialize(resource)
         @resource = resource
         @site = resource.site
-        execute_inline_ruby
-        @output_ext = output_ext_from_converters
+      end
+
+      # @return [String]
+      def output_ext
+        @output_ext ||= output_ext_from_converters
       end
 
       # @return [String]
       def final_ext
+        output_ext # we always need this to get run
+
         permalink_ext || output_ext
       end
 
@@ -33,6 +35,12 @@ module Bridgetown
           run_conversions
           resource.place_in_layout? ? place_into_layouts : resource.output = resource.content.dup
         end
+      end
+
+      def execute_inline_ruby!
+        return unless site.config.should_execute_inline_ruby?
+
+        Bridgetown::Utils::RubyExec.search_data_for_ruby_code(resource, self)
       end
 
       def inspect
@@ -102,12 +110,6 @@ module Bridgetown
 
       ### Transformation Actions
 
-      def execute_inline_ruby
-        return unless site.config.should_execute_inline_ruby?
-
-        Bridgetown::Utils::RubyExec.search_data_for_ruby_code(resource, self)
-      end
-
       def run_conversions # rubocop:disable Metrics/AbcSize
         input = resource.content.to_s
 
@@ -125,7 +127,7 @@ module Bridgetown
             output: Bridgetown.env.production? ? nil : output,
             output_ext: conversions[index][:output_ext],
           }
-          output
+          output.html_safe
         rescue StandardError => e
           Bridgetown.logger.error "Conversion error:",
                                   "#{converter.class} encountered an error while "\