+++ bridgetown-core/lib/bridgetown-core/utils.rb
@@ -5,6 +5,7 @@ module Bridgetown
     extend self
     autoload :Ansi, "bridgetown-core/utils/ansi"
     autoload :RubyExec, "bridgetown-core/utils/ruby_exec"
+    autoload :RubyFrontMatterDSL, "bridgetown-core/utils/ruby_front_matter"
     autoload :Platforms, "bridgetown-core/utils/platforms"
     autoload :ThreadEvent, "bridgetown-core/utils/thread_event"
 
@@ -118,7 +119,13 @@ module Bridgetown
     # @return [Boolean] if the YAML front matter is present.
     # rubocop: disable Naming/PredicateName
     def has_yaml_header?(file)
-      File.open(file, "rb", &:readline).match? %r!\A---\s*\r?\n!
+      File.open(file, "rb", &:readline).match? Bridgetown::FrontMatterImporter::YAML_HEADER
+    rescue EOFError
+      false
+    end
+
+    def has_rbfm_header?(file)
+      File.open(file, "rb", &:readline).match? Bridgetown::FrontMatterImporter::RUBY_HEADER
     rescue EOFError
       false
     end