+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # main
 
+# 0.21.0.beta1 / 2020-04-25
+
+* End-to-end Ruby [front matter, templates](https://beta.bridgetownrb.com/docs/resources#ruby-front-matter-and-all-ruby-templates), and [data files](https://beta.bridgetownrb.com/docs/datafiles) ([jaredcwhite](https://github.com/jaredcwhite)) [#285](https://github.com/bridgetownrb/bridgetown/pull/285)
+* New `Bridgetown::Component` class with a ViewComponent-inspired API [#268](https://github.com/bridgetownrb/bridgetown/pull/268) ([jaredcwhite](https://github.com/jaredcwhite)) [Read the Docs](https://www.bridgetownrb.com/docs/components/ruby)
+  * **Breaking Change:** ERB now uses an output safety buffer to escape HTML in strings, same as in Rails. [Read the Docs](https://beta.bridgetownrb.com/docs/erb-and-beyond#escaping-and-html-safety)
+* Relations for resources (belongs_to, has_many, etc.) [#261](https://github.com/bridgetownrb/bridgetown/pull/261) ([jaredcwhite](https://github.com/jaredcwhite)) [Read the Docs](https://beta.bridgetownrb.com/docs/resources#resource-relations)
+* Migrate to dart-sass since node-sass is deprecated [#279](https://github.com/bridgetownrb/bridgetown/pull/279) ([eclectic-coding](https://github.com/eclectic-coding))
+
 # 0.20.0 / 2020-03-15
 
 * EXPERIMENTAL (and opt-in): the Great Resource Content Engine PR has been merged. 135 files changed. Holy guacamole! [#243](https://github.com/bridgetownrb/bridgetown/pull/243) ([jaredcwhite](https://github.com/jaredcwhite)) [Read the Docs](https://www.bridgetownrb.com/docs/resources)